+++ src/components/v2/V2Project/EditTokenAllocationContent.tsx
@@ -1,5 +1,5 @@
-import { t } from '@lingui/macro'
-import { Modal } from 'antd'
+import { Trans } from '@lingui/macro'
+import { Button } from 'antd'
 import { V2ProjectContext } from 'contexts/v2/projectContext'
 import { useSetProjectSplits } from 'hooks/v2/transactor/SetProjectSplits'
 import { Split } from 'models/v2/splits'
@@ -8,17 +8,9 @@ import { emitErrorNotification } from 'utils/notifications'
 import { preciseFormatSplitPercent } from 'utils/v2/math'
 
 import { RESERVED_TOKEN_SPLIT_GROUP } from 'constants/v2/splits'
-import { EditTokenAllocation } from '../../EditTokenAllocation'
+import { EditTokenAllocation } from './EditTokenAllocation'
 
-export const EditTokenAllocationModal = ({
-  visible,
-  onOk,
-  onCancel,
-}: {
-  visible: boolean
-  onOk: VoidFunction
-  onCancel: VoidFunction
-}) => {
+export const EditTokenAllocationContent = () => {
   const { reservedTokensSplits, fundingCycle } = useContext(V2ProjectContext)
 
   const setProjectSplits = useSetProjectSplits({
@@ -27,7 +19,7 @@ export const EditTokenAllocationModal = ({
 
   const [editingReservedTokensSplits, setEditingReservedTokensSplits] =
     useState<Split[]>([])
-  const [modalLoading, setModalLoading] = useState<boolean>(false)
+  const [loading, setLoading] = useState<boolean>(false)
 
   useEffect(() => {
     if (!reservedTokensSplits) return
@@ -46,7 +38,7 @@ export const EditTokenAllocationModal = ({
 
   const onSaveTokenAllocation = useCallback(async () => {
     if (totalPercentagesInvalid) return
-    setModalLoading(true)
+    setLoading(true)
     const tx = await setProjectSplits(
       {
         groupedSplits: {
@@ -56,39 +48,31 @@ export const EditTokenAllocationModal = ({
       },
       {
         onConfirmed: () => {
-          setModalLoading(false)
-          onOk()
+          setLoading(false)
         },
-        onError: () => setModalLoading(false),
+        onError: () => setLoading(false),
       },
     )
     if (!tx) {
       emitErrorNotification('Token allocation edit failed')
-      setModalLoading(false)
+      setLoading(false)
     }
-  }, [
-    editingReservedTokensSplits,
-    onOk,
-    setProjectSplits,
-    totalPercentagesInvalid,
-  ])
+  }, [editingReservedTokensSplits, setProjectSplits, totalPercentagesInvalid])
 
   return (
-    <Modal
-      visible={visible}
-      confirmLoading={modalLoading}
-      title={t`Edit reserved token allocation`}
-      okText={t`Save token allocation`}
-      cancelText={modalLoading ? t`Close` : t`Cancel`}
-      width={720}
-      onOk={() => onSaveTokenAllocation()}
-      onCancel={onCancel}
-      okButtonProps={{ disabled: totalPercentagesInvalid }}
-    >
+    <>
       <EditTokenAllocation
         editingReservedTokensSplits={editingReservedTokensSplits}
         setEditingReservedTokensSplits={setEditingReservedTokensSplits}
       />
-    </Modal>
+      <Button
+        loading={loading}
+        onClick={() => onSaveTokenAllocation()}
+        disabled={totalPercentagesInvalid}
+        type="primary"
+      >
+        <Trans>Save token allocation</Trans>
+      </Button>
+    </>
   )
 }
