+++ src/components/v2/V2Project/V2ProjectSettings/V2ProjectSettingsReservedTokensContent.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const V2ProjectSettingsReservedTokensContent = () => {
+  return <div>V2ProjectSettingsReservedTokensContent</div>
+}
+
+export default V2ProjectSettingsReservedTokensContent
