+++ src/components/v2/V2Project/V2ProjectSettings/V2ProjectSettingsSidebar.tsx
@@ -1,0 +1,5 @@
+const V2ProjectSettingsSidebar = () => {
+  return <div>V2ProjectSettingsSidebar</div>
+}
+
+export default V2ProjectSettingsSidebar
