+++ src/components/v2/V2Project/V2ProjectSettings/V2ProjectSettingsV1TokenMigrationContent.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const V2ProjectSettingsV1TokenMigrationContent = () => {
+  return <div>V2ProjectSettingsV1TokenMigrationContent</div>
+}
+
+export default V2ProjectSettingsV1TokenMigrationContent
