+++ src/components/v2/V2Project/V2ProjectToolsDrawer/V2ProjectToolsDrawer.tsx
@@ -2,8 +2,6 @@ import { Trans } from '@lingui/macro'
 import { Divider, Drawer, Space, Tabs } from 'antd'
 import { ExportSection } from 'components/Project/ProjectToolsDrawer/ExportSection'
 import ArchiveV2Project from 'components/v2/V2Project/ArchiveV2Project'
-import VeNftEnableSection from 'components/veNft/VeNftEnableSection'
-import VeNftSetUnclaimedTokensPermissionSection from 'components/veNft/VeNftSetUnclaimedTokensPermissionSection'
 import { FEATURE_FLAGS } from 'constants/featureFlags'
 import {
   ETH_PAYOUT_SPLIT_GROUP,
@@ -12,11 +10,9 @@ import {
 import { V2ProjectContext } from 'contexts/v2/projectContext'
 import { useIsUserAddress } from 'hooks/IsUserAddress'
 import useMobile from 'hooks/Mobile'
-import useUserUnclaimedTokenBalance from 'hooks/v2/contractReader/UserUnclaimedTokenBalance'
+import { default as useUserUnclaimedTokenBalance } from 'hooks/v2/contractReader/UserUnclaimedTokenBalance'
 import { useAddToBalanceTx } from 'hooks/v2/transactor/AddToBalanceTx'
 import { useDeployProjectPayerTx } from 'hooks/v2/transactor/DeployProjectPayerTx'
-import { useEditV2ProjectDetailsTx } from 'hooks/v2/transactor/EditV2ProjectDetailsTx'
-import { useTransferProjectOwnershipTx } from 'hooks/v2/transactor/TransferProjectOwnershipTx'
 import { useTransferUnclaimedTokensTx } from 'hooks/v2/transactor/TransferUnclaimedTokensTx'
 import { ETHPayoutSplitGroup, ReservedTokensSplitGroup } from 'models/v2/splits'
 import { useContext } from 'react'
@@ -25,6 +21,7 @@ import { AddToProjectBalanceForm } from '../../../Project/ProjectToolsDrawer/Add
 import { PayableAddressSection } from '../../../Project/ProjectToolsDrawer/PayableAddressSection'
 import { TransferOwnershipForm } from '../../../Project/ProjectToolsDrawer/TransferOwnershipForm'
 import { TransferUnclaimedTokensForm } from '../../../Project/ProjectToolsDrawer/TransferUnclaimedTokensForm'
+import V2ProjectSettingsVenftContent from '../V2ProjectSettings/V2ProjectSettingsVenftContent'
 import { ExportSplitsButton } from './ExportSplitsButton'
 import { V1TokenMigrationSetupSection } from './V1TokenMigrationSetupSection'
 
@@ -42,12 +39,10 @@ export function V2ProjectToolsDrawer({
     tokenSymbol,
     payoutSplits,
     reservedTokensSplits,
-    veNft: { contractAddress: veNftContractAddress },
   } = useContext(V2ProjectContext)
 
   const isMobile = useMobile()
 
-  const editV2ProjectDetailsTx = useEditV2ProjectDetailsTx()
   const { data: unclaimedTokenBalance } = useUserUnclaimedTokenBalance()
 
   const isOwnerWallet = useIsUserAddress(projectOwnerAddress)
@@ -61,25 +56,12 @@ export function V2ProjectToolsDrawer({
       <Divider />
 
       <section>
-        <TransferOwnershipForm
-          ownerAddress={projectOwnerAddress}
-          useTransferProjectOwnershipTx={useTransferProjectOwnershipTx}
-        />
+        <TransferOwnershipForm />
       </section>
 
       <Divider />
 
-      <ArchiveV2Project editV2ProjectDetailsTx={editV2ProjectDetailsTx} />
-    </Space>
-  )
-
-  const VeNftTools = (
-    <Space direction="vertical" size="large" style={{ width: '100%' }}>
-      {!veNftContractAddress ? (
-        <VeNftEnableSection />
-      ) : (
-        <VeNftSetUnclaimedTokensPermissionSection />
-      )}
+      <ArchiveV2Project />
     </Space>
   )
 
@@ -152,7 +134,7 @@ export function V2ProjectToolsDrawer({
         )}
         {veNftEnabled && isOwnerWallet && (
           <TabPane tab={<Trans>veNFT</Trans>} key="3">
-            {VeNftTools}
+            <V2ProjectSettingsVenftContent />
           </TabPane>
         )}
       </Tabs>
