+++ src/utils/featureFlags.ts
@@ -13,6 +13,9 @@ const FEATURE_FLAG_DEFAULTS: {
   [FEATURE_FLAGS.VENFT_CREATOR]: {
     rinkeby: true,
   },
+  [FEATURE_FLAGS.SETTINGS_PAGE]: {
+    rinkeby: true,
+  },
 }
 
 const featureFlagKey = (baseKey: string) => {
